+++ Cargo.toml
@@ -4,7 +4,7 @@ members = [
   "components/spider-core",
   "components/spider-derive",
   "components/spider-execution-manager",
-  "components/spider-proto",
+  "components/spider-proto-rust",
   "components/spider-storage",
   "components/spider-task-executor",
   "components/spider-tdl",
@@ -20,7 +20,7 @@ default-members = [
   "components/spider-core",
   "components/spider-derive",
   "components/spider-execution-manager",
-  "components/spider-proto",
+  "components/spider-proto-rust",
   "components/spider-storage",
   "components/spider-task-executor",
   "components/spider-tdl",
