+++ components/spider-execution-manager/Cargo.toml
@@ -18,7 +18,6 @@ futures-util = {
 }
 rmp-serde = "1.3.1"
 spider-core = { path = "../spider-core" }
-spider-proto = { path = "../spider-proto" }
 spider-task-executor = { path = "../spider-task-executor" }
 spider-tdl = { path = "../spider-tdl" }
 thiserror = "2.0.18"
