+++ components/spider-proto-rust/src/lib.rs
@@ -1,0 +1,6 @@
+//! Rust gRPC protocol definitions generated from Spider protobuf files.
+
+#[allow(clippy::all, clippy::nursery, clippy::pedantic)]
+pub mod storage {
+    include!("generated/storage.rs");
+}
