--- components/spider-proto/src/lib.rs
@@ -1,8 +1,0 @@
-//! gRPC protocol definitions for Spider storage services.
-
-#[allow(clippy::all, clippy::nursery, clippy::pedantic)]
-pub mod generated {
-    include!("generated/spider.storage.rs");
-}
-
-pub mod storage;
